+++ src/Armada.Core/Database/Sqlite/Implementations/CoordinationClaimMethods.cs
@@ -79,7 +79,7 @@ namespace Armada.Core.Database.Sqlite.Implementations
                     using (SqliteDataReader reader = await cmd.ExecuteReaderAsync(token).ConfigureAwait(false))
                     {
                         if (await reader.ReadAsync(token).ConfigureAwait(false))
-                            return SqliteDatabaseDriver.CoordinationClaimFromReader(reader);
+                            return CoordinationClaimColumns.Read(reader, SqliteDatabaseDriver.StoredValues);
                     }
                 }
             }
@@ -154,7 +154,7 @@ namespace Armada.Core.Database.Sqlite.Implementations
                     using (SqliteDataReader reader = await cmd.ExecuteReaderAsync(token).ConfigureAwait(false))
                     {
                         while (await reader.ReadAsync(token).ConfigureAwait(false))
-                            results.Add(SqliteDatabaseDriver.CoordinationClaimFromReader(reader));
+                            results.Add(CoordinationClaimColumns.Read(reader, SqliteDatabaseDriver.StoredValues));
                     }
                 }
             }